requirements.txt
 Flask 
 Flask-Cors 
 pymongo[srv] 
 python-dotenv